% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedVis.R
\name{bedVis}
\alias{bedVis}
\title{bedVis}
\arguments{
\item{bdFile}{the bed file path, default(NULL).}

\item{chr}{the chromesome of peak, default(NULL).}

\item{region.min}{the peak start coordinate, default(NULL).}

\item{region.max}{the peak end coordinate, default(NULL).}

\item{track.width}{track width, default(0.1).}

\item{collapse}{whether collapse the track, default(FALSE).}

\item{fill}{track fill colors, default(NULL).}

\item{show.legend}{whether show fill color legend, default(TRUE).}

\item{add.label}{whether add peak name, default(FALSE).}

\item{label.column}{the peak name column name, default(NULL).}

\item{label.vjsut}{the peak label vjust, default(0.1).}
}
\value{
a ggplot object.
}
\description{
visualize peaks(bed files).
}
\author{
JunZhang
}
