% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcuRotatedCoord.R
\name{calcuRotatedCoord}
\alias{calcuRotatedCoord}
\title{calcuRotatedCoord}
\arguments{
\item{data}{data.frame}

\item{theta}{rotate degree, default(45).}

\item{workers}{how many workers for parallel calculation, default(1).}

\item{rx}{x variable name, default(NULL).}

\item{ry}{y variable name, default(NULL).}
}
\value{
a data.frame
}
\description{
calculate the rotated rectangle coordinate with specified degree.
}
\author{
JunZhang
}
