% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkVis.R
\name{linkVis}
\alias{linkVis}
\title{linkVis}
\usage{
linkVis(
  linkData = NULL,
  start = NULL,
  end = NULL,
  group = NULL,
  base_size = 14,
  link.aescolor = NULL,
  link.color = NULL,
  line.size = 0.5,
  curvature = 0.5,
  yshift = 0.1,
  legend.title = "",
  facet = TRUE,
  facet.placement = "outside",
  facet.fill = "grey90",
  facet.color = "black",
  facet.text.angle = 90,
  facet.text.size = 14,
  xAixs.info = TRUE
)
}
\arguments{
\item{linkData}{the link data with data.frame format, default(NULL).}

\item{start}{link start position, default(NULL).}

\item{end}{link end position, default(NULL).}

\item{group}{facet group variable name, default(NULL).}

\item{base_size}{theme base_size, default(14).}

\item{link.aescolor}{link line color or mapping variable name, default(NULL).}

\item{link.color}{colors to change the link line colors when "link.aescolor" is a mapping variable, default(NULL).}

\item{line.size}{link line size, default(0.5).}

\item{curvature}{the link line curvature, default(0.5).}

\item{yshift}{the space upper the link line, default(0.1).}

\item{legend.title}{the legend title, default("").}

\item{facet}{whether show the plot with facet plot, default(TRUE).}

\item{facet.placement}{the facet label placement, default("outside").}

\item{facet.fill}{facet rectangle fill color, default("grey90").}

\item{facet.color}{facet rectangle border color, default("black").}

\item{facet.text.angle}{facet text angle, default(90).}

\item{facet.text.size}{facet text size, default(14).}

\item{xAixs.info}{whether remove X axis info, default(FASLE).}
}
\value{
a ggplot object.
}
\description{
visualize the coordinate relation like chromtin accessbility or peak sites correlation.
}
\author{
JunZhang
}
