% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackVis.R
\name{trackVis}
\alias{trackVis}
\title{trackVis}
\arguments{
\item{bWData}{the data.frame bigwig data, default(NULL).}

\item{gtf.file}{whether supply gtf annotation file, default(NULL).}

\item{gene.name}{the gene name to be choosed for visualization, default(NULL).}

\item{chr}{chr the chromesome of peak, default(NULL).}

\item{region.min}{the start coordinate, default(NULL).}

\item{region.max}{the end coordinate, default(NULL).}

\item{show.legend}{whether show color legend, default(FALSE).}

\item{legend.position}{the legend position, default("right").}

\item{color}{the track color, default(NULL).}

\item{extend.up}{extend for upstream of start site, default(3000).}

\item{extend.dn}{extend for downstream of start site, default(3000).}

\item{base_size}{theme base size, default(14).}

\item{label.angle}{the facet label angle, default(0).}

\item{label.face}{the facet label face, default("bold").}

\item{space.y}{the facet panel space, default(0.5).}

\item{sample.order}{the sample order to be plotted in graph, default(NULL).}

\item{sampleName.dist}{the facet label distance from Y axis, default(0.4).}

\item{sampleName.hjust}{the facet label hjust, default(1).}

\item{sampleName.vjust}{the facet label vjust, default(0.5).}

\item{xAxis.info}{whether retain X axis info, default(TRUE).}

\item{yAxis.info}{whether retain Y axis info, default(TRUE).}

\item{ticks.y.len}{the y axis ticks length, default(0.3).}

\item{theme}{plot theme, "bw" or "classic", default("classic").}

\item{scales}{the facet scales settings, default("fixed").}

\item{ncol}{the columns to be arranged, default(1).}

\item{mark.region}{whether highlight regions in plot, default(FALSE).}

\item{mark.col}{the colors of marked regions, default(NULL).}

\item{mark.alpha}{the color alpha of marked regions, default(0.5).}

\item{new.yaxis}{whether add new style Y axis, default(FALSE).}

\item{pos.ratio}{the new style Y axis relative position, default(c(0.05,0.8)).}

\item{yinfo.text.size}{the new style Y axis text size, default(5).}
}
\value{
a ggplot object.
}
\description{
visualize bigwig files.
}
\author{
JunZhang
}
